added Object.Generation property and precondition check when uploading.
Server will now return 412 Precondition failed error on uploads if client set the DoesNotExist or Generation condition on the object handle it uploads, like the real server does.

Also adds a new ServerError and function to write the error response with correct StatusCode and reduce code duplication in the various handler functions that write error responses.